Unify docs site, fix CI deploy, and clean up cross-platform code
- Migrate docs CI to peaceiris/actions-gh-pages for external repo deploy
- Redesign docs homepage with updated hero, features grid, and styling
- Update codelabs, comparison, showcase, and v4-preview pages
- Add UI test target to SceneViewDemo Xcode project
- Remove unused import in AssetPath.kt
- Fix TimeSource.ios to use proper Foundation import

+++ sceneview-core/src/iosMain/kotlin/io/github/sceneview/utils/TimeSource.kt
@@ -11,6 +11,6 @@ import kotlinx.cinterop.ptr
 @OptIn(ExperimentalForeignApi::class)
 actual fun nanoTime(): Long = memScoped {
     val ts = alloc<timespec>()
-    clock_gettime(CLOCK_MONOTONIC, ts.ptr)
-    ts.tv_sec * 1_000_000_000L + ts.tv_nsec
+    clock_gettime(CLOCK_MONOTONIC.toUInt(), ts.ptr)
+    ts.tv_sec * 1_000_000_000L + ts.tv_nsec.toLong()
 }
